Do not call async tests in python 2
